Relocate FeatureFormView.NavigationPathItem
